% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeRemovalFunctions.R
\name{dspk.DataGapInterpolation}
\alias{dspk.DataGapInterpolation}
\title{Linear interpolation}
\usage{
dspk.DataGapInterpolation(
  Data = NULL,
  Value,
  precision = NULL,
  NumDateTime = NULL,
  max.gap = Inf,
  State.of.value.data = NULL,
  state.of.value.code = 93,
  NAvalue = NULL,
  logoutput = F
)
}
\arguments{
\item{precision}{If you enter in a precision then it will round the interpolated values. Precision is the is the smallest measurable unit on the scale. e.g. 13000 would have precision = 1000 and 0.23 would have precision = 0.01}

\item{max.gap}{is the largest data gap that you want to perform linear interpolation on.}

\item{logoutput}{}
}
\value{

}
\description{
Linear interpolation
}
