% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeRemovalFunctions.R
\name{dspk.getmode}
\alias{dspk.getmode}
\title{Mode}
\usage{
dspk.getmode(v)
}
\arguments{
\item{v}{A vector. May be a string, numeric, ect.}
}
\value{
The mode of the input vector
}
\description{
Mode
}
