% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeRemovalFunctions.R
\name{dspk.TableFormatting}
\alias{dspk.TableFormatting}
\title{Pre-format data tables for use with other HICbioclean functions}
\usage{
dspk.TableFormatting(
  Data = NULL,
  Value,
  DateTime = NULL,
  NAvalue = NULL,
  datetime.format = NULL,
  datetime.timezone = "GMT",
  state.of.value.code = 110,
  state.of.value.code.na = 255,
  add.original.data = T
)
}
\arguments{
\item{Data}{a datatable or NULL}

\item{Value}{a vector, column number or column name}

\item{DateTime}{a vector, column number, column name or NULL. If left NULL then a time will be generated by numbering the records consecutivly.}

\item{NAvalue}{NULL or the value to be turned to NA}

\item{datetime.format}{If date time is in character format then a format must be given so it can be converted to POSIX format.  Character string giving the datetime format. See strptime() documentation for additional help.}

\item{datetime.timezone}{Character string giving the time zone of the datetime. By default “GMT”. See strptime() documentation for additional help.}

\item{state.of.value.code}{State of value to give all non-NA values}

\item{state.of.value.code.na}{State of value to give all NA values}

\item{add.original.data}{If TRUE then all original data in the provided datatable will be included in the output data table}
}
\value{
data table with values in "dspk.Values", numeric datetime in "dspk.DateTimeNum", and state of value in "dspk.StateOfValue".
if add.original.data is TRUE then all the data in the input data table will be included.
}
\description{
This function is not exported
}
\details{
Puts data values and datetime in a unique column and creates a state of value column. If the state of value column "dspk.StateOfValue" already
exists then those values will be kept as the state of values. if datetime is numeric then those values will be kept. If date time is
a character format, then it will be converted into UNIX seconds. If datetime is missing then the records will be numbered cosecutivly.
}
