% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUICodless.R
\name{HIC.App.manual}
\alias{HIC.App.manual}
\title{Graphical applications for formatting, auto-validation, manual-validation and final export of HIC database data}
\usage{
HIC.App.manual()
}
\value{
Three shiny apps
}
\description{
Three graphical user interfaces (GUI) for the processing of the Flemish Hydrological Information Center (HIC) data.
These graphical apps will walk you through the entire work flow of data import, validation/calibration and data export, in an intuitive visual manner without the need for coding.
This is the only method of manual validation and calibration and final export.
}
\details{
Either enter "format" to format HIC database data, "auto" to auto-validate formatted data, or "manual" to manually check and calibrate auto-validated data against reference site locations (if available) and export final zrx files for upload back into the HIC database.

The formatting step and the auto validation step can be done in R code using the functions HIC.Continuous.Data.Import.Format(), spk.DespikingWorkflow.CSVfileBatchProcess(), and HIC.PPFDAutoValidation.CSVfileBatchProcess().
The manual validation, calibration and final export must be done using this graphical app.

\subsection{Recomended workflow}
First: format the csv files that are exported from the HIC database - HIC.App.format()

Second: auto-validate the formatted files - HIC.App.auto()

Third: manual-validate/calibrate the files and export the data - HIC.App.manual()

Files can be batch processed at each step so there is no need to run through all the steps for each individual file.

\subsection{Manual Tutorial}
\url{https://github.com/pgelsomini/HICbioclean/blob/master/MANUAL-TUTORIAL-HICbioclean.-Rpackage.pdf}
}
\examples{
HIC.App.format() # to format the HIC database output csv files
HIC.App.auto() # to auto-validate the formatted data
HIC.App.manual() # to manually check the auto-validated data
}
