% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadinXLSfiles.R
\name{HIC.maint}
\alias{HIC.maint}
\title{Batch process folder for converting maintenance Excel files to csv files}
\usage{
HIC.maint(input.directory, output.directory)
}
\arguments{
\item{input.directory}{folder directory of Excel files. No back slashes(\), only use forward slashes(/) or double back slashes(\\\\).}

\item{output.directory}{folder directory where you would like to save the csv files.  No back slashes(\), only use forward slashes(/) or double back slashes(\\\\).}
}
\value{
A folder of csv files with the same name as the input excel files.
}
\description{
This function was specifically made to convert the excel sensor maintenance files into csv files that can be easily read into R.
}
\details{
It batch processes all the files inside a given folder.

If you don't provide a full path to the output directory then it will be placed in your working directory.
}
\examples{
HIC.maint("C:/Rdata/MaintenanceFiles", "MaintenanceFilesCSV")
}
