% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HICWebservicesBioDataDownload.R
\name{HICwebservicesBioDownload}
\alias{HICwebservicesBioDownload}
\title{Function for downloading all continuous biological
data from the HIC server from a given year}
\usage{
HICwebservicesBioDownload(
  year,
  credentials = NULL,
  other.group.ids = NULL,
  output.dir = NULL
)
}
\arguments{
\item{year}{The year that you wish to download}

\item{credentials}{This is the credentials input variable from the HICwebservices package function get_token(). See documentation from HICwebservices package.}

\item{other.group.ids}{Any other timeseries group IDs that you wish to also download. Take note that the entire group will be downloaded.}

\item{output.dir}{The folder name where you wish all the downloaded files to be saved to. If left as NULL it will by default save to 'DownloadedHICDataYYYYMMDDHHMMSS'}
}
\value{
A folder with all the downloaded time series as csv files
and two additional csv files (DownloadSummaryBinary.csv and
DownloadSummaryID.csv) that give the summary of which files were downloaded and site coordinates in Belgian Lambert 72.
The metadata columns DateTimeUnix, Parameter.Name, Station.Name, Station.Number,
Parameter.Unit are added to the downloaded table.
}
\description{
Downloads all the HIC data for Chfyla, DO, pH and PPFD for a given year from group id 446639. Additional
time series groups can also be added to the download list.
The package HICwebservices is required.
}
