% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ReferenceSiteLinkage}
\alias{ReferenceSiteLinkage}
\title{Continuous measurement sites and their reference sites.}
\format{
A data frame with 2 variables:
\describe{
  \item{MonthlySites}{Site names of the nearest periodically measured site}
  \item{ContinuousSites}{Site numbers of the continuously measured sites}
  ...
}
}
\source{
Flemish Hydrological Information Center, OMES Monitoring
}
\usage{
ReferenceSiteLinkage
}
\description{
Dataset with the site numbers of the continuously measured biological parameters and their nearest periodically measured reference site.
}
\keyword{datasets}
