% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zrxFileStationCodes}
\alias{zrxFileStationCodes}
\title{Import codes for the Flemish Hydrological Information Center database .}
\format{
A data frame with 4 variables:
\describe{
  \item{StationNo}{Continuous measurement station number}
  \item{Par}{Parameter}
  \item{Code}{Database code}
  \item{Unit}{Unit}
  ...
}
}
\source{
Flemish Hydrological Information Center
}
\usage{
zrxFileStationCodes
}
\description{
Codes for each site number and parameter for importation into the Flemish Hydrological Information Center database.
}
\keyword{datasets}
