% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeRemovalFunctions.R
\name{dspk.MinMaxfilter}
\alias{dspk.MinMaxfilter}
\title{Min max filter}
\usage{
dspk.MinMaxfilter(
  Data = NULL,
  Value,
  Min,
  Max,
  State.of.value.data = NULL,
  state.of.value.code = 91,
  NAvalue = NULL,
  logoutput = F
)
}
\arguments{
\item{logoutput}{}
}
\value{
returns $data which is a datatable with columns dspk.Values and dspk.StateOfValue containing the filtered data.
If logoutput is TRUE, $logdata contains the info for the log file
}
\description{
Min max filter
}
