% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeRemovalFunctions.R
\name{dspk.Spikefilter}
\alias{dspk.Spikefilter}
\title{Spike removal algorithm}
\usage{
dspk.Spikefilter(
  Data = NULL,
  Value,
  NumDateTime = NULL,
  sampling.interval = NULL,
  State.of.value.data = NULL,
  state.of.value.code = 92,
  good.state.of.value.code = 80,
  NAvalue = NULL,
  threshhold = 3,
  Method = "median",
  logoutput = F
)
}
\arguments{
\item{logoutput}{}
}
\value{
returns $data which is a datatable with columns dspk.Values and dspk.StateOfValue containing the filtered data.
If logoutput is TRUE, $logdata contains the info for the log file
}
\description{
Spike removal algorithm
}
